Feature/upgrade nx (#12)
* feat: Add flight header component and update flight booking UI

This commit adds the flight header component to the flight booking UI. It also updates the flight booking UI by renaming the ProfileUiPage component to ProfileUI. Additionally, the storybook link and code tab UI have been updated.

#9

* feat: Add flight card preview component to flight booking page

This commit adds the flight card preview component to the flight booking page, allowing users to see a preview of a flight card. The flight card preview component is integrated with the existing flight booking UI and provides an onPress event handler to handle click events on the flight card.

* feat: Add flight card preview component and flight header component to flight booking UI

This commit adds the flight card preview component and the flight header component to the flight booking UI. The flight card preview component allows users to see a preview of a flight card, and it is integrated with the existing flight booking UI. The flight header component provides additional functionality to the flight booking UI, including the ability to hide the header on scroll. This commit also includes updates to the storybook link and code tab UI.

#9

* feat: Refactor flight card component and update flight booking UI

This commit refactors the flight card component to improve code organization and readability. It also updates the flight booking UI by renaming the ProfileUiPage component to ProfileUI. Additionally, the storybook link and code tab UI have been updated.

* Refactor flight card component and update flight booking UI

* feat: Add flight list preview component to flight booking UI

* feat: Add flight filter form preview component to flight booking UI

* feat: Update flight filter form import in UI components

This commit updates the import statement for the flight filter form component in the UI components file. The import was changed from './forms' to './form-fields' to reflect the updated file name. This change ensures that the correct component is imported and used in the UI.

#9

* feat: Update flight filter form import in UI components

* Refactor flight card component and update flight booking UI

* - Upgrade NX

* Refactor package.json script for building Storybook